Restore non-directory-based output format for v0.3 (#42)
## Background

We have big plans to update the output format for commec, so that outputs are organized into folders with the following hierarchy:

* {queryname}
  * {queryname}.screen.log
  * {queryname}.json
  * {queryname}.html
   * {queryname}_input
       * _dir of generated input files, e.g. {queryname}.cleaned.fasta_
   * {queryname}_output
       * _dir of intermediate output files, e.g. {queryname}.nt.blastn_

We did a version of this in https://github.com/ibbis-bio/common-mechanism/pull/30/, but realized that this was too breaking of a change for a minor update (e.g. `commec flag` made assumptions about the output structure, we were a little too aggressive in our approach to shortening filenames).

As a result, this undoes that change (which we will likely reintroduce in v1.0) and adds a dev-script, `collate-screens.py`, which allows reformatting outputs from v0.3 to a flags-ingestible format.

## Changes

### Bug fixes
* Expand variables and avoid creating directories named "~"

### New features
* Restore previous output format, including handling of directory inputs
* Add collate-screens dev script