Allow trailing slashes for base_paths, fix check_biorisk return code bug (#45)
Addresses two minor changes that we wanted to have in 0.3.1 concerning
the -d CLI command, and the taxonomy yaml directory structure.

The base path used to update the internal config dict is modified to
always contain "/" at the end. This affects both the base path in the
yaml, as well as fixing the -d CLI version.

The yaml taxonomy dictionaries are now updated to be located within the
databases section. On top of this, the yaml taxonomy defined
dictionaries are actually used in the code as well. (Rather than legacy
hard-coded guesses)

* Correctly assembles path when base_paths missing trailing "/"
* Addresses bug reported in
https://github.com/ibbis-bio/common-mechanism/issues/46

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>
Co-authored-by: Tessa Alexanian <hello@tessa.fyi>