Formatless logging (#56)
## Background
Sometimes, you need to copy paste a url, or console command from debug
mode, and the beautiful formatting for the logs was making that
issuesome. Furthermore, the final table summary print at the end of a
screen run could benefit from not being beholden to the formatting
system used in logging. Thus we add a way for the formatting to be
removed, and altered, from the system currently in place.

## Changes
Any logger function can now parse some keywords within the extra
dictionary, which allows custom logging behaviour - mainly to remove
suffix log formatting , and optionally neatly cap off the existing
formatting.

## Features
All console outputs which feature a some copy/paste convenient output,
are logged on a single line, with no maximum width, to make copy/pasting
easy for the user.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>
Co-authored-by: Tessa Alexanian <alexanian@users.noreply.github.com>