Adds additional debug output to record input CLI and parameters used in screen. (#60)
## Background
We had an issue with -force result in -f -o rce, this may of been more
quickly identified with more verbose debug logging. This PR aims to
improve future argument and parameter visibility in a debug context.

## Changes
Prints the args.NameSpace and the screen parameters dictionary overidden
by yaml after settings parsing to the debug log.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>
Co-authored-by: Tessa Alexanian <alexanian@users.noreply.github.com>