Rationale (#57)
## Background
Whilst looking at a list of FLAG, WARN, and PASS screen status' is nice,
what is missing is the rationale behind these decisions in a human
readable format. This adds an initial implementation of rationale to the
recommendation container of a query, and gives a quick grab text, from
`screendata.query.status.rationale`, for providing the user or
tester to read.

## Changes
* A rational text is determined during screen data update.
* Screen data update logic has been improved to be more robust.
* Benign steps are no longer always PASS, and reflect the status of what
they did. e.g. Flag (Cleared)
* The summary table is now printed in two, with a table of outcomes for each step,
and a table of rationales for each overall outcome.
* The HTML has been updated to include the rationale in the headings,
and in the hover over text of the query bar.
* Each query has a new boolean flag, `no_hits_warning`, initialised to
`True`, as commec screen runs, the query is updated to check for whether
it has hit _anything_ at all. When updating screen data, if a query is
still True, it will marked as a warning (instead of PASS from all steps)
and the rationale being there is no known homology.
** Note: a future implementation can build on this behaviour and detect
poorly represented regions rather than treat the query as a whole.
* Sequences that are too short, as dictated by the new constant, are rejected,
and skipped as a whole from the commec process. Their rationale text reflects this.
 
### Bug fixes
* Fixed an off-by-one graphic glitch in HTML bar rendering.

### New features
* Rationale is added in Commec Outputs.
* Short queries are skipped gracefully.
---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>
Co-authored-by: Tessa Alexanian <alexanian@users.noreply.github.com>