Setup hotfix and DB location support (#66)
## Background
`commec setup` has a minor typo error which breaks nt database download.
This fixes issues #65.
Furthermore, commec specific databases are now hosted in the
commec-databases repository, and this now points to the latest vX.X.X
database from that. Note this updates dynamically depending on the most
recent tagged release in commec databases repo.

**Issues**:

Fixes issue #65 

## Changes
Changes the default url for biorisk and non-concern (transitionally
benign) databases. The default URL is now updated according to what the
most recent database is from the repository. If this fails, then no
default is offered, and the usual process of the user entering in their
own URL applies.

### Bug fixes
* Fixes a typo resulting in a crash when downloading the nt databases.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>