Legacy flag patch (#73)
## Background
There was duplicate code in legacy flag, and this led to issues in
maintaining the function of this script.

## Changes
legacy flag now references the json import scripts from flag, to ensure
it is always up to date with the latest json imports. The .screen import
script is subtly changed to ensure empty lines dont result in an
out-of-scope declaration of the lines tokens.

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>