Database validation override (#77)
## Background
Adds a simply database override statement such that if `-R, --resume` is
used, and the expected output already exists (i.e. the database handler
would not run the search anyway, then the database files themselves are
not checked for existence / validity.

This allows the complete run of a commec screen based on pre-existing
output alone - with no large external database file setup
[biorisk/nr/nt/benign].

This is useful for testing, reproducibility, and tutorial purposes.

Note, we still require correct config yaml imports of the annotations
and taxid files for biorisk and benign - these are however very
lightweight, and could be bundled with a yaml easily for tutorial output
purposes.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>
Co-authored-by: Tessa Alexanian <hello@tessa.fyi>