Fixes Diamond empty file outputs. (#74)
## Background
Unlike BlastX, and the other tools, Diamond output files contain no
header information. Due to this, the output checks were conflating the
logic of "file doesn't exists" with "file exists, but is empty" when
determining whether to proceed with Step 2. This was leading to
erroneous ERROR statuses being present when Diamond was returning no
hits.


**Issues**:
Addresses issue #63 


## Changes
The `check_output()` method is now, `validate_output()` and _only_ used, instead of _as well as_
the `is_empty()` method to determine whether the output of a handler is
valid. for further STEP processes. Importantly, Diamond has overridden
the method to ignore if the file is empty, but does exist, which is a
valid output for Diamond Handlers.

To ensure that the `has_hits`, and `is_empty` is used correctly, the
static method status for each has been removed. This has resulted in a
refactor of `calculate_noncoding_regions_per_query` to take Query
dictionaries and the handler, rather than arbitrary file names. `is_empty` is renamed to `has_no_output()` to align it with other similar function names.

Furthermore, old unused legacy based code for historic fetching of nc
bits has been removed, and the associated test rewritten to incorporate
the used methods, as well as correctly skipping a < 50 bp non-coding
multi-query output.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>
Co-authored-by: Tessa Alexanian <hello@tessa.fyi>