Include Synthetic Constructs in query has hits logic (#76)
## Background
get_taxonomic_labels currently filters out synthetic constructs when
adding the labels. This means that when checking taxonomic results,
those hits were being excluded when checking whether or not the query
had any hits at all. This was leading to some queries which were wholly
synthetic constructs being mis-labelled as Warnings, due to being
mysteriously lacking in any know taxonomic hit.

## Changes
A check for hits is performed on the blast outputs directly, before any
filtering takes place. (Rather than just on the top hits of
non-synthetic constructs). Any hits and a query can
be confirmed it has a hit, and is therefore not an unidentified
sequence at this early stage.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>