Query name fix (#96)
## Background
An issue appeared when trying to run the
sample_NIST-screening-dataset-blinded-seqs.fa examples, where the length
of the names were too long and were not being processed correctly for
internal ID use. This PR addresses that issue and provides a fix.

## Changes
Query names that are not tokenable (i.e. no not have underscores) and
are longer than 64 characters are simply truncated to 64 characters.

Raised Errors with the text from same-id queries are correctly passed to
the user.

Extended query id from 25 to 64.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>