Collapse Taxonomic hits based on hit name, rather than the hit accession string. (#97)
The hit name is more likely than accession to not be unique, providing a
better user experience for parsing verbose repeated like sequences, so
we now collapse hits based on name.

The functional test outputs were carefully curated to ensure that the
result was the same, and required changes to some hit names to ensure
that the benign checks treated appropriate regions separately. Also
minor updates to the checks for testing regions in screen.

Whilst consider the name or accession as a collapse point, this created
complications in checking, as the accession is not typically stored to
be checked. In either case, the accession tends to be more unique, and
more often the name is shared with unique accessions, rather than the
other way around.

---------

Co-authored-by: Michael Barnett <ToMichaelBarnett@gmail.com>